+++ crates/nostr-database/src/error.rs
@@ -10,13 +10,10 @@ use thiserror::Error;
 #[derive(Debug, Error)]
 pub enum DatabaseError {
     /// An error happened in the underlying database backend.
-    #[error("backend: {0}")]
+    #[error(transparent)]
     Backend(Box<dyn std::error::Error + Send + Sync>),
-    /// Nostr error
-    #[error("nostr: {0}")]
-    Nostr(Box<dyn std::error::Error + Send + Sync>),
     /// Not supported
-    #[error("method not supported by current backend")]
+    #[error("not supported by current backend")]
     NotSupported,
     /// Feature disabled
     #[error("feature disabled for current backend")]
@@ -33,16 +30,5 @@ impl DatabaseError {
         E: std::error::Error + Send + Sync + 'static,
     {
         Self::Backend(Box::new(error))
-    }
-
-    /// Create a new [`Nostr`][Self::Nostr] error.
-    ///
-    /// Shorthand for `Error::Nostr(Box::new(error))`.
-    #[inline]
-    pub fn nostr<E>(error: E) -> Self
-    where
-        E: std::error::Error + Send + Sync + 'static,
-    {
-        Self::Nostr(Box::new(error))
     }
 }